refactor: move environment actions to dedicated route files
