+++ packages/insomnia/src/ui/routes/$organizationId.project.$projectId.workspace.$workspaceId.environment.create.tsx
@@ -1,0 +1,24 @@
+import { type ActionFunctionArgs } from 'react-router';
+
+import * as models from '../../models';
+import { EnvironmentType } from '../../models/environment';
+import { invariant } from '../../utils/invariant';
+
+export async function action({ request, params }: ActionFunctionArgs) {
+  const { workspaceId } = params;
+  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
+
+  const { isPrivate, environmentType = EnvironmentType.KVPAIR } = await request.json();
+
+  const baseEnvironment = await models.environment.getByParentId(workspaceId);
+
+  invariant(baseEnvironment, 'Base environment not found');
+
+  const environment = await models.environment.create({
+    parentId: baseEnvironment._id,
+    environmentType,
+    isPrivate,
+  });
+
+  return environment;
+}
