+++ packages/insomnia/src/ui/routes/$organizationId.project.$projectId.workspace.$workspaceId.environment.delete.tsx
@@ -1,0 +1,26 @@
+import { type ActionFunctionArgs } from 'react-router';
+
+import * as models from '../../models';
+import { invariant } from '../../utils/invariant';
+
+export async function action({ request, params }: ActionFunctionArgs) {
+  const { workspaceId } = params;
+  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
+
+  const formData = await request.formData();
+
+  const environmentId = formData.get('environmentId');
+  invariant(typeof environmentId === 'string', 'Environment ID is required');
+
+  const environment = await models.environment.getById(environmentId);
+
+  const baseEnvironment = await models.environment.getByParentId(workspaceId);
+
+  invariant(environment?._id !== baseEnvironment?._id, 'Cannot delete base environment');
+
+  invariant(environment, 'Environment not found');
+
+  await models.environment.remove(environment);
+
+  return null;
+}
