+++ packages/insomnia/src/ui/routes/$organizationId.project.$projectId.workspace.$workspaceId.environment.set-active.tsx
@@ -1,0 +1,23 @@
+import { type ActionFunctionArgs } from 'react-router';
+
+import * as models from '../../models';
+import { invariant } from '../../utils/invariant';
+
+export async function action({ request, params }: ActionFunctionArgs) {
+  const { workspaceId } = params;
+  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
+
+  const formData = await request.formData();
+
+  const environmentId = formData.get('environmentId');
+
+  invariant(typeof environmentId === 'string', 'Environment ID is required');
+
+  const workspaceMeta = await models.workspaceMeta.getOrCreateByParentId(workspaceId);
+
+  invariant(workspaceMeta, 'Workspace meta not found');
+
+  await models.workspaceMeta.update(workspaceMeta, { activeEnvironmentId: environmentId || null });
+
+  return null;
+}
