+++ packages/insomnia/src/ui/routes/$organizationId.project.$projectId.workspace.$workspaceId.environment.update.tsx
@@ -1,0 +1,26 @@
+import { type ActionFunctionArgs } from 'react-router';
+
+import * as models from '../../models';
+import { invariant } from '../../utils/invariant';
+
+export async function action({ request, params }: ActionFunctionArgs) {
+  const { workspaceId } = params;
+  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
+
+  const { environmentId, patch } = await request.json();
+
+  invariant(typeof environmentId === 'string', 'Environment ID is required');
+
+  const environment = await models.environment.getById(environmentId);
+
+  invariant(environment, 'Environment not found');
+  invariant(typeof name === 'string', 'Name is required');
+
+  const baseEnvironment = await models.environment.getByParentId(workspaceId);
+
+  invariant(baseEnvironment, 'Base environment not found');
+
+  const updatedEnvironment = await models.environment.update(environment, patch);
+
+  return updatedEnvironment;
+}
